Commit up to Lesson 22 (Final Lesson)
Brushed up on form instancing and ownership which I'm already familiar with; passing variables.

Leaving the assignment up to this lesson since I'm already familiar with MSSQL.

There was a bug where CellClick events would happen twice when selling and buying.
This may be caused as the button triggers the event in addition to focusing cell click which triggers it a second time. Temporary Solution was to add a boolean flag to stop the second trigger and reset itself.